Fix `StepperItemPropso` definition place
